-fix cal methond :float to double

+++ include/cal_method.h
@@ -5,7 +5,9 @@
 #include <math.h>
 #include <stdint.h>
 #include <string.h>
+#include <stdlib.h>
 #include "Tree.h"
+
 
 #ifndef PI
     #define PI acos(-1)
@@ -14,7 +16,7 @@
 
 typedef struct Trace_info
 {
-    float entropy;
+    double entropy;
     uint32_t total_count;
     uint32_t distinct ;
 }trace_info_t;
